using static cast to make the uint -> int conversion explicit and prevent a compiler warning

+++ src/io/bluetooth/BluetoothDeviceHandler.cpp
@@ -41,7 +41,7 @@ namespace zen
 
     nonstd::expected<std::vector<std::byte>, ZenError> BluetoothDeviceHandler::read() noexcept
     {
-        auto nBytes = m_device->Read(reinterpret_cast<char *>(m_buffer.data()), m_buffer.size());
+        auto nBytes = m_device->Read(reinterpret_cast<char *>(m_buffer.data()), static_cast<int>(m_buffer.size()));
         SPDLOG_DEBUG("Read {0} bytes trom bluetooth interface", nBytes);
         if (nBytes == 0)
             return nonstd::make_unexpected(ZenError_Io_ReadFailed);
